Medical Appointments Predictive Model
Being able to minimise the rate of patients not showing up for an allocated appointment is of paramount interest to any medical practice. Unused capacity and misallocation of resources arise as a result of the no- shows, as the booked time slots can be used more efficiently. In order to reduce the monetary losses associated with this problem, it is necessary to understand the behavior behind no-shows, and be able to identify and predict which patients are associated with not showing up. With an accurate prediction model, the medical practitioners would be able to intervene in advance and better mitigate the risks of no shows.
For this case, we have picked a medical practice in Brazil. Our goal is to identify which characteristics are strongly associated with no-show behavior and to be able to predict the probability of a no-show for any given patient. With the outcome of our prediction model, we can provide the upper management with the best recommendations for which measures to implement to reduce the costs of no-shows in the future. More specifically, we will strive to offer our client guidance in three parts:
Investigate general patterns and provide information about the patients.
Explore relationships between no show behaviour and the characteristics of the patients. Build a prediction model, that can accurately predict the likelihood of patients not showing up.